requirements.txt
aiogram==3.4.1
openai==1.6.1
python-dotenv==1.0.0
aiohttp==3.9.1
aiosqlite==0.19.0
reportlab==4.0.9
aioschedule==0.5.2